requirements.txt
python-telegram-bot==21.6
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
